+++ server/ghc-persistent-worker-server.cabal
@@ -17,6 +17,7 @@ executable Server
     ghc-options:      -Wall -Werror
     other-modules:    Pool
                       Server
+                      Worker
     hs-source-dirs:   app
     build-depends:    base >=4.19,
                       binary,
@@ -30,5 +31,6 @@ executable Server
                       optparse-applicative,
                       process,
                       stm,
+                      transformers,
                       unix
     default-language: GHC2021
